fix: update modals delete to overlay
